# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/friend/agree": {
            "post": {
                "tags": [
                    "好友"
                ],
                "summary": "同意好友请求",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "被同意的用户id",
                        "name": "agreedId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/friend/delete": {
            "post": {
                "tags": [
                    "好友"
                ],
                "summary": "删除好友",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "被删除的用户id",
                        "name": "deletedId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/friend/getList": {
            "get": {
                "tags": [
                    "好友"
                ],
                "summary": "获取好友列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
        "/friend/request": {
            "post": {
                "tags": [
                    "好友"
                ],
                "summary": "发送好友请求",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authenticate",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "被请求的用户id",
                        "name": "requestedId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "401": {
                        "description": "验证失败",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/server.RespJson"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.RespJson": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                },
                "data": {},
                "msg": {}
            }
        }
    }
}